fix(backend): patch security and logic bugs in course, booking, mentor-availability
- Fix unauthenticated JWT decode auth bypass in GET /courses
- Fix isFree filter applied after pagination causing incorrect meta.total
- Add ownership check to MentorAvailabilityService.update()
- Fix missing CANCEL enum value on mentor_availabilities.status
- Restrict booking status transition to COMPLETED to prevent fake reviews

feat(frontend): add SEO infrastructure and fix course/calendar UX bugs

- Add robots.ts, sitemap.ts, unify SITE_URL, add JSON-LD and per-page metadata
- Refactor course detail and mentor profile pages to server components for metadata
- Fix booking/calendar timezone mismatch, wire up checkConflict, remove fake rating fallback
