+++ backend/package.json
@@ -22,6 +22,7 @@
     "hygen": "hygen",
     "gen:module": "node scripts/gen-module.js",
     "seed:categories": "ts-node scripts/seed-categories.ts",
+    "sync:category-icons": "ts-node scripts/update-category-icons.ts",
     "seed:admin": "ts-node scripts/seed-admin.ts",
     "seed:courses": "ts-node scripts/seed-courses.ts",
     "seed:mentor-courses": "ts-node scripts/seed-custom-mentor-courses.ts"
