+++ backend/src/modules/course/course.module.ts
@@ -1,5 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { JwtModule } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
 import { CourseController } from './course.controller';
 import { CourseService } from './services/course.service';
 import { CourseEntity } from './entities/course.entity';
@@ -12,6 +14,13 @@ import { CourseCategoryModule } from '../course-category/course-category.module'
   imports: [
     TypeOrmModule.forFeature([CourseEntity, CourseCategoryEntity]),
     CourseCategoryModule,
+    JwtModule.registerAsync({
+      imports: [ConfigModule],
+      useFactory: async (configService: ConfigService) => ({
+        secret: configService.get<string>('jwt.accessSecret'),
+      }),
+      inject: [ConfigService],
+    }),
   ],
   controllers: [CourseController],
   providers: [CourseService, CourseRepository, CourseResolver],
